swift: add Core integration tests
